made some progress on the http crud, messages are now working the way i want them to. just have to refactor the react and maybe replace the message when someone likes it so i can dispaly the total likes. will require a bit of backend for that as well. not too much left to go - pick up with finishing likes then move on to friends and requests then finish with private messages before moving onto sockets and webrtc

+++ backend/src/routes/services/like.service.ts
@@ -3,6 +3,24 @@ import { prisma } from '../../config/database';
 import type { Like } from '@prisma/client';
 
 /**
+ * Get users liked messages
+ * @param userId - the user id
+ * @returns the messages ids
+ */
+export const getLikedMessages = async (userId: number): Promise<number[]> => {
+  const likes = await prisma.like.findMany({
+    where: {
+      userId,
+    },
+    select: {
+      messageId: true,
+    },
+  });
+
+  return likes.map((like: Like) => like.messageId);
+};
+
+/**
  * Create a new like
  * @param userId - the user id
  * @param messageId - the message id
@@ -32,13 +50,28 @@ export const createLike = async (
 
 /**
  * Delete a like
- * @param likeId - the like id
+ * @param userId - the user id
+ * @param messageId - the message id
  * @returns the like
  * */
-export const deleteLike = async (likeId: number): Promise<Like> => {
-  const like = await prisma.like.delete({
+export const deleteLike = async (
+  userId: number,
+  messageId: number,
+): Promise<Like> => {
+  const like = await prisma.like.findFirst({
     where: {
-      id: likeId,
+      userId,
+      messageId,
+    },
+  });
+
+  if (!like) {
+    throw new Error('Like not found');
+  }
+
+  await prisma.like.delete({
+    where: {
+      id: like.id,
     },
   });
